Add database seed data and routes
The Courses and Holes tables now have seeder data. This seeder data is
in the format of an exported javascript object. Adding more seeder data
to this file is a simple matter of adding the appropriate rows.

The courses seeder data is in the format of:

[COURSE_NAME,NUMBER_OF_HOLES,COURSE_PAR,ADDRESS]

The holes seeder data is in the format of:

[HOLE_NUMBER,COURSE_FOREIGN_KEY,PAR]

At the present moment, there are 3 routes:
- `/api/courses`
- `/api/courses/:id`
- `/api/courses/:id/holes`.

The first route returns a JSON object that contains all data from all
courses. The second route returns data for a single course. The third
route returns all data about a given course's holes.